Added a way of checking if the person sending the messages is the manager Added vscode exceptions to the gitignore.

+++ bot/cogs/message_filter.py
@@ -15,7 +15,7 @@ class MessageFilter(commands.Cog):
     async def on_message(self, message):
         if message.guild is None or message.author == self.bot.user:
             return
-        if message.channel.id == 720394587792867378 or message.author == self.bot.user:
+        if message.channel.id == 720394587792867378 or message.author == self.bot.user and not await bot_global.is_bot_manager(message.author):
             if message.content != "!agree":
                 await message.delete()
                 return
